+++ tests/conftest.py
@@ -3,5 +3,5 @@ Module py.test fixtures
 """
 # pylint: disable=unused-import
 
-from .fixtures.backends import es, es_data_stream, swift  # noqa: F401
+from .fixtures.backends import es, es_data_stream, events, swift, ws  # noqa: F401
 from .fixtures.logs import gelf_logger  # noqa: F401